+++ src/sorbetExtensionContext.ts
@@ -3,7 +3,6 @@ import { Configuration } from './common/configuration';
 import { Log } from './common/log';
 import { LogMetrics, Metrics } from './common/metrics';
 import { SorbetClientManager } from './lspClient/sorbetClientManager';
-import { SorbetStatusProvider } from './sorbetStatusProvider';
 
 export class SorbetExtensionContext implements vscode.Disposable {
   public readonly clientManager: SorbetClientManager;
@@ -12,7 +11,6 @@ export class SorbetExtensionContext implements vscode.Disposable {
   public readonly extensionContext: vscode.ExtensionContext;
   public readonly logOutputChannel: vscode.LogOutputChannel;
   public readonly metrics: Metrics;
-  public readonly statusProvider: SorbetStatusProvider;
 
   constructor(context: vscode.ExtensionContext) {
     this.configuration = new Configuration();
@@ -21,14 +19,8 @@ export class SorbetExtensionContext implements vscode.Disposable {
 
     this.clientManager = new SorbetClientManager(this);
     this.metrics = new LogMetrics(this.logOutputChannel);
-    this.statusProvider = new SorbetStatusProvider(this);
 
-    this.disposables = [
-      this.configuration,
-      this.clientManager,
-      this.logOutputChannel,
-      this.statusProvider,
-    ];
+    this.disposables = [this.configuration, this.clientManager, this.logOutputChannel];
   }
 
   dispose(): void {