+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## 0.3.6
-- Improve tracking of context editors.
+- Improved tracking of context by looking at additional types of editors.
+- Fix: Tracking of Sorbet operations not multi-root aware.
 - Fix: Tracking of active editor from 0.3.5 was incomplete.
 
 ## 0.3.5